# serviflex-server/docs/swagger.yaml
definitions:
...
  models.Agendamento:
    properties:
      cliente_id:
        type: string
      data_hora:
        type: string
      id:
        type: string
      procedimento:
        type: string
      profissional_id:
        type: string
    type: object
...
  models.Procedimento:
    properties:
      descricao:
        type: string
      duracao_min:
        type: integer
      id:
        type: string
      imagem_url:
        type: string
      nome:
        type: string
      preco:
        type: number
      profissional_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /agendamentos:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do agendamento
        in: body
        name: agendamento
        required: true
        schema:
          $ref: '#/definitions/models.Agendamento'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Agendamento'
      summary: Agendar horário
      tags:
      - Cliente
  /agendamentos/cliente/{id}:
    get:
      parameters:
      - description: ID do cliente
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Agendamento'
            type: array
      summary: Listar agendamentos do cliente
      tags:
      - Agendamentos
...
  /procedimentos:
    post:
      consumes:
      - application/json
      parameters:
      - description: Procedimento
        in: body
        name: procedimento
        required: true
        schema:
          $ref: '#/definitions/models.Procedimento'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Procedimento'
      summary: Criar procedimento
      tags:
      - Procedimentos
  /procedimentos/{id}:
    delete:
      parameters:
      - description: ID do procedimento
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deletar procedimento
      tags:
      - Procedimentos
    get:
      parameters:
      - description: ID do profissional
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Procedimento'
            type: array
      summary: Listar procedimentos por profissional
      tags:
      - Procedimentos
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do procedimento
        in: path
        name: id
        required: true
        type: string
      - description: Procedimento atualizado
        in: body
        name: procedimento
        required: true
        schema:
          $ref: '#/definitions/models.Procedimento'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar procedimento
      tags:
      - Procedimentos
...
swagger: "2.0"